chore: move scheduler ui to features dir
