+++ packages/frontend/src/components/SchedulersView/SchedulersViewActionMenu.tsx
@@ -1,3 +1,4 @@
+import { SchedulerDeleteModal } from '@features/scheduler';
 import { ActionIcon, Menu } from '@mantine/core';
 import {
     IconDots,
@@ -9,7 +10,6 @@ import React, { FC } from 'react';
 import { useQueryClient } from 'react-query';
 import { Link } from 'react-router-dom';
 import MantineIcon from '../common/MantineIcon';
-import SchedulerDeleteModal from '../SchedulerModals/SchedulerModalBase/SchedulerDeleteModal';
 import {
     getItemLink,
     getSchedulerLink,
