+++ packages/frontend/src/features/scheduler/components/SchedulerDeleteModal.tsx
@@ -7,17 +7,17 @@ import {
     NonIdealState,
     Spinner,
 } from '@blueprintjs/core';
+import ErrorState from '@ui/common/ErrorState';
 import React, { FC, useCallback, useEffect } from 'react';
-import { useScheduler } from '../../../hooks/scheduler/useScheduler';
-import { useSchedulersDeleteMutation } from '../../../hooks/scheduler/useSchedulersDeleteMutation';
-import ErrorState from '../../common/ErrorState';
+import { useScheduler } from '../hooks/useScheduler';
+import { useSchedulersDeleteMutation } from '../hooks/useSchedulersDeleteMutation';
 
 interface DashboardDeleteModalProps extends DialogProps {
     schedulerUuid: string;
     onConfirm: () => void;
 }
 
-const SchedulerDeleteModal: FC<DashboardDeleteModalProps> = ({
+export const SchedulerDeleteModal: FC<DashboardDeleteModalProps> = ({
     schedulerUuid,
     onConfirm,
     ...modalProps
@@ -81,5 +81,3 @@ const SchedulerDeleteModal: FC<DashboardDeleteModalProps> = ({
         </Dialog>
     );
 };
-
-export default SchedulerDeleteModal;
