+++ packages/frontend/src/features/scheduler/components/SchedulersList.tsx
@@ -1,9 +1,9 @@
 import { NonIdealState, Spinner } from '@blueprintjs/core';
 import { ApiError, SchedulerAndTargets } from '@lightdash/common';
+import ErrorState from '@ui/common/ErrorState';
 import React, { FC, useState } from 'react';
 import { UseQueryResult } from 'react-query/types/react/types';
-import ErrorState from '../../common/ErrorState';
-import SchedulerDeleteModal from './SchedulerDeleteModal';
+import { SchedulerDeleteModal } from './SchedulerDeleteModal';
 import SchedulersListItem from './SchedulersListItem';
 
 type Props = {
