+++ packages/frontend/src/features/scheduler/hooks/useSchedulersUpdateMutation.ts
@@ -4,8 +4,8 @@ import {
     UpdateSchedulerAndTargetsWithoutId,
 } from '@lightdash/common';
 import { useMutation, useQueryClient } from 'react-query';
-import { lightdashApi } from '../../api';
-import useToaster from '../toaster/useToaster';
+import { lightdashApi } from '../../../api';
+import useToaster from '../../../hooks/toaster/useToaster';
 
 const updateScheduler = async (
     uuid: string,
