+++ packages/frontend/src/features/sync/components/SyncModal.tsx
@@ -1,10 +1,10 @@
+import { getSchedulerUuidFromUrlParams } from '@features/scheduler/utils';
 import { Flex, Modal, ModalProps, Title } from '@mantine/core';
 import { IconTrash } from '@tabler/icons-react';
+import { GSheetsIcon } from '@ui/common/GSheetsIcon';
 import { FC, useEffect } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { GsheetsIcon } from '../../../components/SchedulerModals/SchedulerModalBase/SchedulerModalBase.styles';
-import { getSchedulerUuidFromUrlParams } from '../../../components/SchedulerModals/SchedulerModalBase/SchedulerModalContent';
 import {
     SyncModalAction,
     SyncModalProvider,
@@ -32,7 +32,7 @@ const SyncModalBaseAndManager: FC<Props> = ({ chartUuid, opened, onClose }) => {
     }, [history, pathname, search, setAction, setCurrentSchedulerUuid]);
 
     let modalTitle = 'Sync with Google Sheets';
-    let headerIcon: typeof GsheetsIcon | typeof IconTrash = GsheetsIcon;
+    let headerIcon: typeof GSheetsIcon | typeof IconTrash = GSheetsIcon;
     let headerIconColor = 'black';
 
     if (action === SyncModalAction.CREATE) {
