+++ packages/frontend/src/features/sync/components/SyncModalForm.tsx
@@ -1,4 +1,7 @@
 import { NonIdealState, Spinner } from '@blueprintjs/core';
+import { useChartSchedulerCreateMutation } from '@features/scheduler/hooks/useChartSchedulers';
+import { useScheduler } from '@features/scheduler/hooks/useScheduler';
+import { useSchedulersUpdateMutation } from '@features/scheduler/hooks/useSchedulersUpdateMutation';
 import {
     CreateSchedulerAndTargetsWithoutIds,
     SchedulerFormat,
@@ -19,9 +22,6 @@ import { FormProvider, useForm } from 'react-hook-form';
 import ErrorState from '../../../components/common/ErrorState';
 import MantineIcon from '../../../components/common/MantineIcon';
 import CronInput from '../../../components/ReactHookForm/CronInput';
-import { useChartSchedulerCreateMutation } from '../../../hooks/scheduler/useChartSchedulers';
-import { useScheduler } from '../../../hooks/scheduler/useScheduler';
-import { useSchedulersUpdateMutation } from '../../../hooks/scheduler/useSchedulersUpdateMutation';
 import { isInvalidCronExpression } from '../../../utils/fieldValidators';
 import { SyncModalAction, useSyncModal } from '../providers/SyncModalProvider';
 import { SelectGoogleSheetButton } from './SelectGoogleSheetButton';
