+++ packages/frontend/src/features/sync/components/SyncModalView.tsx
@@ -1,3 +1,4 @@
+import { useChartSchedulers } from '@features/scheduler/hooks/useChartSchedulers';
 import { SchedulerFormat } from '@lightdash/common';
 import {
     ActionIcon,
@@ -14,7 +15,6 @@ import { IconDots, IconPencil, IconTrash } from '@tabler/icons-react';
 import cronstrue from 'cronstrue';
 import { FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { useChartSchedulers } from '../../../hooks/scheduler/useChartSchedulers';
 import { SyncModalAction, useSyncModal } from '../providers/SyncModalProvider';
 
 export const SyncModalView: FC<{ chartUuid: string }> = ({ chartUuid }) => {
