+++ packages/frontend/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "extends": "./../../tsconfig.json",
     "compilerOptions": {
+        "baseUrl": "./src",
         "composite": true,
         "target": "ESNext",
         "lib": ["dom", "dom.iterable", "esnext"],
@@ -21,7 +22,10 @@
         "jsx": "react-jsx",
         "types": ["jest", "node", "vite/client", "vite-plugin-svgr/client"],
         "paths": {
-            "@lightdash/common": ["./../common/src"]
+            "@lightdash/common": ["./../common/src"],
+            "@features/*": ["features/*"],
+            // Alias shared components to @ui
+            "@ui/*": ["components/*"]
         }
     },
     "include": ["src"]
